+++ numpy/core/src/multiarray/_multiarray_tests.c.src
@@ -1867,7 +1867,7 @@ static PyMethodDef Multiarray_TestsMethods[] = {
 #if defined(NPY_PY3K)
 static struct PyModuleDef moduledef = {
         PyModuleDef_HEAD_INIT,
-        "multiarray_tests",
+        "_multiarray_tests",
         NULL,
         -1,
         Multiarray_TestsMethods,
@@ -1880,11 +1880,11 @@ static struct PyModuleDef moduledef = {
 
 #if defined(NPY_PY3K)
 #define RETVAL m
-PyMODINIT_FUNC PyInit_multiarray_tests(void)
+PyMODINIT_FUNC PyInit__multiarray_tests(void)
 #else
 #define RETVAL
 PyMODINIT_FUNC
-initmultiarray_tests(void)
+init_multiarray_tests(void)
 #endif
 {
     PyObject *m;
@@ -1892,7 +1892,7 @@ initmultiarray_tests(void)
 #if defined(NPY_PY3K)
     m = PyModule_Create(&moduledef);
 #else
-    m = Py_InitModule("multiarray_tests", Multiarray_TestsMethods);
+    m = Py_InitModule("_multiarray_tests", Multiarray_TestsMethods);
 #endif
     if (m == NULL) {
         return RETVAL;
@@ -1900,7 +1900,7 @@ initmultiarray_tests(void)
     import_array();
     if (PyErr_Occurred()) {
         PyErr_SetString(PyExc_RuntimeError,
-                        "cannot load umath_tests module.");
+                        "cannot load _multiarray_tests module.");
     }
     return RETVAL;
 }
