+++ numpy/core/src/umath/_rational_tests.c.src
@@ -1129,7 +1129,7 @@ PyMethodDef module_methods[] = {
 #if defined(NPY_PY3K)
 static struct PyModuleDef moduledef = {
     PyModuleDef_HEAD_INIT,
-    "test_rational",
+    "_rational_tests",
     NULL,
     -1,
     module_methods,
@@ -1142,10 +1142,10 @@ static struct PyModuleDef moduledef = {
 
 #if defined(NPY_PY3K)
 #define RETVAL m
-PyMODINIT_FUNC PyInit_test_rational(void) {
+PyMODINIT_FUNC PyInit__rational_tests(void) {
 #else
 #define RETVAL
-PyMODINIT_FUNC inittest_rational(void) {
+PyMODINIT_FUNC init_rational_tests(void) {
 #endif
 
     PyObject *m = NULL;
@@ -1295,7 +1295,7 @@ PyMODINIT_FUNC inittest_rational(void) {
 #if defined(NPY_PY3K)
     m = PyModule_Create(&moduledef);
 #else
-    m = Py_InitModule("test_rational", module_methods);
+    m = Py_InitModule("_rational_tests", module_methods);
 #endif
 
     if (!m) {
@@ -1397,7 +1397,7 @@ PyMODINIT_FUNC inittest_rational(void) {
 fail:
     if (!PyErr_Occurred()) {
         PyErr_SetString(PyExc_RuntimeError,
-                        "cannot load test_rational module.");
+                        "cannot load _rational_tests module.");
     }
 #if defined(NPY_PY3K)
     if (m) {
