+++ numpy/core/src/umath/_umath_tests.c.src
@@ -360,7 +360,7 @@ static PyMethodDef UMath_TestsMethods[] = {
 #if defined(NPY_PY3K)
 static struct PyModuleDef moduledef = {
         PyModuleDef_HEAD_INIT,
-        "umath_tests",
+        "_umath_tests",
         NULL,
         -1,
         UMath_TestsMethods,
@@ -373,11 +373,11 @@ static struct PyModuleDef moduledef = {
 
 #if defined(NPY_PY3K)
 #define RETVAL m
-PyMODINIT_FUNC PyInit_umath_tests(void)
+PyMODINIT_FUNC PyInit__umath_tests(void)
 #else
 #define RETVAL
 PyMODINIT_FUNC
-initumath_tests(void)
+init_umath_tests(void)
 #endif
 {
     PyObject *m;
@@ -387,7 +387,7 @@ initumath_tests(void)
 #if defined(NPY_PY3K)
     m = PyModule_Create(&moduledef);
 #else
-    m = Py_InitModule("umath_tests", UMath_TestsMethods);
+    m = Py_InitModule("_umath_tests", UMath_TestsMethods);
 #endif
     if (m == NULL)
         return RETVAL;
@@ -406,7 +406,7 @@ initumath_tests(void)
 
     if (PyErr_Occurred()) {
         PyErr_SetString(PyExc_RuntimeError,
-                        "cannot load umath_tests module.");
+                        "cannot load _umath_tests module.");
     }
 
     return RETVAL;
